+++ resources/views/pages/about.blade.php
@@ -1,0 +1,15 @@
+<x-app-layout>
+    <x-slot name="header">
+        <h2 class="font-semibold text-xl text-gray-800 leading-tight">
+            {{ __('About us') }}
+        </h2>
+    </x-slot>
+
+    <div class="py-2">
+        <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
+            <div class="py-8">
+                {!! config('settings.about') !!}
+            </div>
+        </div>
+    </div>
+</x-app-layout>