+++ resources/views/partials/footer.blade.php
@@ -3,31 +3,35 @@
     <div class="max-w-7xl mx-auto py-12 px-4 overflow-hidden sm:px-6 lg:px-8">
         <nav class="-mx-5 -my-2 flex flex-wrap justify-center" aria-label="Footer">
             <div class="px-5 py-2">
-                <a href="#" class="text-base text-gray-500 hover:text-gray-900">
+                <a href="/"
+                    class="text-base text-gray-500 hover:text-gray-900 dark:text-gray-300 dark:hover:text-gray-500">
+                    Home
+                </a>
+            </div>
+
+            <div class="px-5 py-2">
+                <a href="{{ route('about')}}"
+                    class="text-base text-gray-500 hover:text-gray-900 dark:text-gray-300 dark:hover:text-gray-500">
                     About
                 </a>
             </div>
 
             <div class="px-5 py-2">
-                <a href="#" class="text-base text-gray-500 hover:text-gray-900">
-                    Blog
+                <a href="{{ route('terms') }}"
+                    class="text-base text-gray-500 hover:text-gray-900 dark:text-gray-300 dark:hover:text-gray-500">
+                    Terms and conditions
                 </a>
             </div>
 
             <div class="px-5 py-2">
-                <a href="#" class="text-base text-gray-500 hover:text-gray-900">
-                    Jobs
+                <a href="{{ route('policy') }}"
+                    class="text-base text-gray-500 hover:text-gray-900 dark:text-gray-300 dark:hover:text-gray-500">
+                    Privacy policy
                 </a>
             </div>
 
-            <div class="px-5 py-2">
-                <a href="#" class="text-base text-gray-500 hover:text-gray-900">
-                    Press
-                </a>
-            </div>
-
-            <div class="px-5 py-2">
-                <a href="#" class="text-base text-gray-500 hover:text-gray-900">
+            {{-- <div class="px-5 py-2">
+                <a href="#" class="text-base text-gray-500 hover:text-gray-900 dark:text-gray-300 dark:hover:text-gray-500">
                     Accessibility
                 </a>
             </div>
@@ -36,7 +40,7 @@
                 <a href="#" class="text-base text-gray-500 hover:text-gray-900">
                     Partners
                 </a>
-            </div>
+            </div> --}}
         </nav>
         <div class="mt-8 flex justify-center space-x-6">
             @if (config('settings.facebook'))